// SDKs/squareup.com/2.0/typescript/src/sdk/models/shared/card.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import { Address } from "./address";


// Card
/** 
 * Represents the payment details of a card to be used for payments. These
 * details are determined by the payment token generated by Web Payments SDK.
**/
export class Card extends SpeakeasyBase {
  @Metadata({ data: "json, name=billing_address" })
  billingAddress?: Address;

  @Metadata({ data: "json, name=bin" })
  bin?: string;

  @Metadata({ data: "json, name=card_brand" })
  cardBrand?: string;

  @Metadata({ data: "json, name=card_type" })
  cardType?: string;

  @Metadata({ data: "json, name=cardholder_name" })
  cardholderName?: string;

  @Metadata({ data: "json, name=customer_id" })
  customerId?: string;

  @Metadata({ data: "json, name=enabled" })
  enabled?: boolean;

  @Metadata({ data: "json, name=exp_month" })
  expMonth?: number;

  @Metadata({ data: "json, name=exp_year" })
  expYear?: number;

  @Metadata({ data: "json, name=fingerprint" })
  fingerprint?: string;

  @Metadata({ data: "json, name=id" })
  id?: string;

  @Metadata({ data: "json, name=last_4" })
  last4?: string;

  @Metadata({ data: "json, name=prepaid_type" })
  prepaidType?: string;

  @Metadata({ data: "json, name=reference_id" })
  referenceId?: string;

  @Metadata({ data: "json, name=version" })
  version?: number;
}